agents/central_agent/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
requests
python-dotenv
pydantic
deepseek 